requirements.txt
torch==1.10.0
transformers==4.12.5
tqdm~=4.64.0
black~=22.3.0
psycopg2-binary==2.9.3
mysql-connector-python==8.0.29